requirements.txt
faiss-cpu
sentence-transformers
langchain
pypdf
langchain_community
streamlit
google-generativeai
